requirements.txt
asgiref==3.6.0
crispy-bootstrap4==2022.1
Django==4.1.7
django-bootstrap-datepicker-plus==5.0.3
django-crispy-forms==2.0
django-jazzmin==2.6.0
mysqlclient==2.1.1
pydantic==1.10.6
sqlparse==0.4.3
typing_extensions==4.5.0
tzdata==2022.7
